fixed heart and moved it to image component
